requirements.txt


Flask==2.0.2
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
pyrsistent==0.19.3
SQLAlchemy==1.4.25
Werkzeug==2.3.6
email-validator==1.1.3
Flask-Principal==0.4.0
Flask-WTF==1.0.0